+++ src/foundry/context.js
@@ -2,6 +2,7 @@ export const context = {
     FilePicker,
     Hooks,
     fetch: fetch.bind(window),
+    fromUuid: fromUuid.bind(window),
 
     get foundry() { return foundry; },
     get game() { return game; },